requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
Django==2.1.7
django-cors-headers==2.5.2
djangorestframework==3.9.2
docopt==0.6.2
dukpy==0.2.2
echarts-china-cities-pypkg==0.0.9
echarts-china-counties-pypkg==0.0.2
echarts-china-misc-pypkg==0.0.1
echarts-china-provinces-pypkg==0.0.3
echarts-countries-pypkg==0.1.6
echarts-united-kingdom-pypkg==0.0.1
fake-useragent==0.1.11
pip
future==0.17.1
gerapy==0.8.5
get==2018.11.19
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
javascripthon==0.10
Jinja2==2.10
jupyter-echarts-pypkg==0.1.2
lml==0.0.2
lxml==4.3.2
macropy3==1.1.0b2
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
parsel==1.5.1
Pillow==5.4.1
post==2018.11.20
protobuf==3.7.0
public==2018.11.20
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
pyecharts==0.5.11
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pyecharts-snapshot==0.1.10
PyHamcrest==1.9.0
pymongo==3.7.2
PyMySQL==0.9.3
pyOpenSSL==19.0.0
python-scrapyd-api==2.1.2
pytz==2018.9
query-string==2018.11.20
queuelib==1.5.0
redis==3.2.1
request==2018.11.20
requests==2.21.0
Scrapy==1.6.0
scrapy-redis==0.6.8
scrapy-splash==0.7.2
scrapyd==1.2.0
scrapyd-client==1.1.0
service-identity==18.1.0
six==1.12.0
Twisted==18.9.0
urllib3==1.24.1
w3lib==1.20.0
zope.interface==4.6.0
